refactor: clean up unused imports and improve code quality
remove unused imports across multiple modules
simplify range check in job_matcher/update.rs
use first() instead of get(0) for better readability
clean up redundant code in job_matcher/pallet.rs

+++ src/backend/src/user_history/tests.rs
@@ -1,5 +1,3 @@
-use super::types::Rating;
-use candid::Principal;
 
 #[test]
 fn test_calic() {
@@ -13,8 +11,8 @@ fn test_calic() {
     let total_rate: Vec<Rating> = vec![rating.clone(), rating.clone()];
     let total_actions_rate: Vec<Rating> = vec![];
 
-    let total_rate_sum: f64 = total_rate.iter().map(|r| r.rating as f64).sum();
-    let total_actions_rate_sum: f64 = total_actions_rate.iter().map(|r| r.rating as f64).sum();
+    let total_rate_sum: f64 = total_rate.iter().map(|r| r.rating).sum();
+    let total_actions_rate_sum: f64 = total_actions_rate.iter().map(|r| r.rating).sum();
 
     let others_rate = total_rate_sum / total_rate.len() as f64;
